android-Camera2Video: fix bug in Camera2Device app on Android11/Pixel4a
When stopping the recording, the correct sequence is to pause the
MediaRecorder, then closing the Recording Session, then closing
and releasing the MediaRecorder.

If the MediaRecorder is released before the Recording Session is
paused, the graphic buffers produced by Camera have nowhere to go,
and seem to not be properly cleaned up by the system,
Android11/Pixel4a device gets in a bad state.